Bugfix: Make CAN scheduler definitions visible.
Internal structure definitions need to be in a separate module so we can
make them visible in the handlers. I don't quite understand why; perhaps
this is a bug in the semantics of Tower?

As long as I was creating a module, I also moved the helper functions
that don't depend on a specific instance of the scheduler global data
structures into the new module.