CAN.Sched: Fix a possible abort race condition.
Removing the aborting task from the queue in the per-task abort handler,
separately from the common task abort handler actually handling the
abort, makes it difficult to reason about whether all invariants are
maintained between handler invocations.

I think in this case it was OK, but with this patch I'm confident,
because the per-task abort handler now has no effect on the scheduler's
monitor state.