CAN.Sched: Fix spurious aborts.
Found by inspection. I don't think this would have been an observable
bug except in unnecessary delay for some outgoing messages, because the
spuriously-aborted request would still be the correct next request and
would just get queued on the mailbox again. But let's not allow the
driver to be that stupid.