tower-freertos-stm32: remove unnecessary systemInit thread
We were unconditionally adding an empty InitThread in compileTowerSTM32FreeRTOS
that was creating a problem later when constructing the graph. Removing it
doesn't appear to create a problem and it seems unnecessary, as the user would
have provided an init if they need one and otherwise init shouldn't be missed.