rewrite i2c driver
This thing is extra fiddly now, but hopefully handles all the special
cases in the datasheet. Notably when receiving, there are three
special cases for receiving n > 2, n = 2, and n = 1 bytes. The
structure of the driver is now a state machine that matches the
synthesis of prose and figures in §27.3.3 of the datasheet.

There is more work that can be done on error handling in exceptional
cases. I've noted in comments states that should be impossible to
reach, but to fully handle them we need to implement a robust reset
that actually reinitializes the peripheral after resetting it with
rcc. TODO if we need it.